restructure: use an array of strings to store lines
Previously the file's data was stored in one large string, but now it is
stored in an array of strings to make it easier to manipulate data. As a
result of this, all rpos and lpos variables have been removed in favor
of vpos. Additionally, the fixrpos method has been removed.

Signed-off-by: Ricky Ayoub <rayoub@redhat.com>