fix: newlines removed, selecting works, print cursor
This patch removes newlines entirely to the array of strings and writes
the \n character while the file is being written to disk. This is done
in such a manner such that it shouldn't interfere with the original line
endings.

The selection cursor has been fixed to work with the new array mode, as
it wasn't in the last commit.

The cursor has also been designed to stay at the end of a written word
once it's inserted. This means that there shouldn't be a case anymore
where the cursor goes "outside" of the currently written text with the
exception of when the document is empty. Appending to the end of a line
is done vim style, by moving to the end (move $) and then using the
append command instead of insert (a or append).

Signed-off-by: Ricky Ayoub <rayoub@redhat.com>