add moving up and down, and creating and joining lines
With this patch, using up and down commands make it possible to move up
and down lines within the file that is being edited. Joining and
creating commands can also be used to create a new linein the file.

As, in the previous commit, it was mentioned that it isn't possible to
select the newline character, in order to actually create a brand
newline in a file, one has to execute a new line word followed by an
append word, such as "newline append". This is necessary as "append
newline" would mean typing "newline" at the point after the cursor.

Vim does not run into this issue, as it is possible to type "enter" in
Vim, whereas in a chat based program "enter" sends the message.

Signed-off-by: Ricky Ayoub <vgmoose@gmail.com>