add find command to locate a specific term in the document
Find vocabulary can now be used to situation the cursor at the position
of the query, much like Vim.

Refind vocabulary is also added to quickly repeat the last find search
that was executed.

Finding will locate the next instance of the search term at the current
position + 1. it will loop around if the search hits the end of the
document.

Signed-off-by: Ricky Ayoub <rayoub@redhat.com>